+++ CHANGELOG.md
@@ -20,7 +20,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Require filter directives such as `@whereKey` in `@delete`, `@forceDelete` and `@restore` https://github.com/nuwave/lighthouse/pull/2289
 - Subscriptions can now be filtered via `$subscriber->socket_id` and `request()->header('X-Socket-ID')` https://github.com/nuwave/lighthouse/pull/2298
 - Make pagination argument `first` non-nullable
-- Include the field `@complexity` is defined upon in its calculation, increasing it by 1
+- Calculate complexity for `@paginate` according to the number of items requested
+- Require argument `resolver` in directive `@complexity`
 - Move interface `Nuwave\Lighthouse\Support\Contracts\GlobalId` to `Nuwave\Lighthouse\GlobalId\GlobalId`
 - Use union type for `$id` argument in `NodeDirective`
 - Replace `Arr::first(array_keys())` with `array_key_first()` in `OrderByDirective`